Django/requirements.txt
django==3.1.5
djangorestframework==3.12.2
django-rest-knox==4.1.0
